// prisma/generated/zod/inputTypeSchemas/WorkSchedulesUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client';
import { z } from 'zod';

export const WorkSchedulesUncheckedCreateInputSchema: z.ZodType<Prisma.WorkSchedulesUncheckedCreateInput> = z.object({
  id: z.number().int().optional(),
  createdAt: z.coerce.date().optional(),
  updatedAt: z.coerce.date().optional(),
  published: z.boolean().optional(),
  days: z.string(),
  from: z.coerce.date(),
  to: z.coerce.date(),
  restaurantId: z.number().int().optional().nullable()
}).strict();

export default WorkSchedulesUncheckedCreateInputSchema;